fix: slug error
Issue #32
