+++ scripts/data_generator.log
@@ -738,3 +738,11 @@ INFO:root:Starting synthetic data generation for horizon: 1M
 INFO:root:Found 16 CSV files: ['/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2014.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2015.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2017.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2016.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2021_upto_31dec2021.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2013.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2020.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2021.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2021_upto_30apr2021.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2013.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2020_to_jun30.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2023_2024.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2023.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2018.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2024_jan_to_oct.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2019.csv']
 INFO:root:Found 68 unique symbols.
 INFO:root:Finished synthetic data generation.
+INFO:root:Starting synthetic data generation for horizon: 1M
+INFO:root:Found 16 CSV files: ['/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2014.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2015.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2017.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2016.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2021_upto_31dec2021.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2013.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2020.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2021.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2021_upto_30apr2021.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2013.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2020_to_jun30.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2023_2024.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2023.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2018.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2024_jan_to_oct.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2019.csv']
+INFO:root:Found 68 unique symbols.
+INFO:root:Finished synthetic data generation.
+INFO:root:Starting synthetic data generation for horizon: 1M
+INFO:root:Found 16 CSV files: ['/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2014.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2015.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2017.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2016.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2021_upto_31dec2021.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2013.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2020.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2021.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2021_upto_30apr2021.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2013.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2020_to_jun30.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_stock_market_sectors_2023_2024.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2023.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2018.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2024_jan_to_oct.csv', '/Users/reinhard/portfolio-optimization-system/scripts/../ml/datasets/NSE_data_all_stocks_2019.csv']
+INFO:root:Found 68 unique symbols.
+INFO:root:Finished synthetic data generation.
